abstracted fallback color logic into a utility file for reusable imports on categories and labels
